+++ tests/tab-size-setting.spec.js
@@ -21,3 +21,12 @@ test("test custom tab size", async ({ page }) => {
     await page.locator("body").press("Tab")
     expect(await heynotePage.getBlockContent(0)).toBe("  ")
 })
+
+test("test indent type", async ({ page }) => {
+    await page.locator("css=.status-block.settings").click()
+    await page.locator("css=li.tab-editing").click()
+    await page.locator("css=select.indent-type").selectOption("tab")
+    await page.locator("body").press("Escape")
+    await page.locator("body").press("Tab")
+    expect(await heynotePage.getBlockContent(0)).toBe("\t")
+})